requirements.txt
chart_studio==1.1.0
matplotlib==3.8.2
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
plotly==5.18.0
python_dateutil==2.8.2
seaborn==0.13.1
streamlit==1.29.0
xlwings==0.30.13
yfinance==0.2.35

